fixes and stuff
