+++ dev-packages/e2e-tests/test-applications/create-remix-app-v2/app/routes/db-ioredis.tsx
