+++ dev-packages/e2e-tests/test-applications/create-remix-app-v2/package.json
@@ -1,6 +1,7 @@
 {
   "private": true,
   "sideEffects": false,
+  "type": "module",
   "scripts": {
     "build": "remix vite:build && pnpm typecheck",
     "dev": "remix vite:dev",
@@ -8,15 +9,20 @@
     "typecheck": "tsc",
     "clean": "npx rimraf node_modules pnpm-lock.yaml",
     "test:build": "pnpm install && pnpm build",
-    "test:assert": "pnpm playwright test"
+    "test:assert": "pnpm playwright test",
+    "test:build:orchestrion": "INJECT_ORCHESTRION=true pnpm test:build",
+    "test:assert:orchestrion": "INJECT_ORCHESTRION=true pnpm test:assert"
   },
   "dependencies": {
     "@sentry/remix": "file:../../packed/sentry-remix-packed.tgz",
+    "@sentry/server-utils": "file:../../packed/sentry-server-utils-packed.tgz",
     "@remix-run/css-bundle": "2.17.4",
     "@remix-run/node": "2.17.4",
     "@remix-run/react": "2.17.4",
     "@remix-run/serve": "2.17.4",
     "isbot": "^3.6.8",
+    "ioredis": "5.10.1",
+    "mysql": "^2.18.1",
     "react": "^18.2.0",
     "react-dom": "^18.2.0"
   },
@@ -25,6 +31,7 @@
     "@sentry-internal/test-utils": "link:../../../test-utils",
     "@remix-run/dev": "2.17.4",
     "@remix-run/eslint-config": "2.17.4",
+    "@types/mysql": "^2.15.26",
     "@types/react": "^18.2.64",
     "@types/react-dom": "^18.2.34",
     "@types/prop-types": "15.7.7",
@@ -35,6 +42,15 @@
   },
   "resolutions": {
     "@types/react": "18.2.22"
+  },
+  "sentryTest": {
+    "variants": [
+      {
+        "build-command": "pnpm test:build:orchestrion",
+        "assert-command": "pnpm test:assert:orchestrion",
+        "label": "create-remix-app-v2 (orchestrion)"
+      }
+    ]
   },
   "volta": {
     "extends": "../../package.json"
