--- dev-packages/e2e-tests/test-applications/create-remix-app-v2/remix.config.js
@@ -1,9 +1,0 @@
-/** @type {import('@remix-run/dev').AppConfig} */
-module.exports = {
-  ignoredRouteFiles: ['**/.*'],
-  // appDirectory: 'app',
-  // assetsBuildDirectory: 'public/build',
-  // serverBuildPath: 'build/index.js',
-  // publicPath: '/build/',
-  serverModuleFormat: 'cjs',
-};
