--- dev-packages/e2e-tests/test-applications/remix-orchestrion/.gitignore
@@ -1,10 +1,0 @@
-node_modules
-
-/.cache
-/build
-/public/build
-.env
-
-/test-results/
-/playwright-report/
-/playwright/.cache/
