--- dev-packages/e2e-tests/test-applications/remix-orchestrion/app/routes/_index.tsx
@@ -1,3 +1,0 @@
-export default function Index() {
-  return <div>home</div>;
-}
