--- dev-packages/e2e-tests/test-applications/remix-orchestrion/start-event-proxy.mjs
@@ -1,6 +1,0 @@
-import { startEventProxyServer } from '@sentry-internal/test-utils';
-
-startEventProxyServer({
-  port: 3031,
-  proxyServerName: 'remix-orchestrion',
-});
