--- dev-packages/e2e-tests/test-applications/remix-orchestrion/tsconfig.json
@@ -1,25 +1,0 @@
-{
-  "include": ["remix.env.d.ts", "./app/**/*.ts", "./app/**/*.tsx"],
-  "exclude": ["node_modules", "build"],
-  "compilerOptions": {
-    "lib": ["DOM", "DOM.Iterable", "ES2019"],
-    "isolatedModules": true,
-    "esModuleInterop": true,
-    "jsx": "react-jsx",
-    "moduleResolution": "bundler",
-    "resolveJsonModule": true,
-    "target": "ES2019",
-    "strict": true,
-    "allowJs": true,
-    "skipLibCheck": true,
-    "forceConsistentCasingInFileNames": true,
-    "baseUrl": ".",
-    "paths": {
-      "~/*": ["./app/*"]
-    },
-    "types": ["@remix-run/node", "vite/client"],
-
-    // Remix takes care of building everything in `remix vite:build`.
-    "noEmit": true
-  }
-}
